Made some changes in the XML of register
