+++ app/src/main/java/com/example/nutritionapp/Register/FragmentEmail.java
@@ -1,16 +1,20 @@
 package com.example.nutritionapp.Register;
 
+import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentTransaction;
 
+import android.util.Patterns;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.EditText;
+import android.widget.Toast;
 
+import com.example.nutritionapp.Activity.Login;
 import com.example.nutritionapp.R;
 import com.example.nutritionapp.databinding.FragmentEmailBinding;
 import com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton;
@@ -19,7 +23,7 @@ import com.google.android.material.floatingactionbutton.FloatingActionButton;
 public class FragmentEmail extends Fragment {
     private FragmentEmailBinding binding;
     RegisterMain registerMain;
-    EditText textEmail, textMobile;
+    EditText textEmail, textMobile,textName;
     ExtendedFloatingActionButton add,back;
 
     @Nullable
@@ -31,8 +35,10 @@ public class FragmentEmail extends Fragment {
 
         textEmail= root.findViewById(R.id.text_email);
         textMobile= root.findViewById(R.id.text_mobile);
+        textName=root.findViewById(R.id.text_name);
         add= root.findViewById(R.id.email_next);
-
+        back=root.findViewById(R.id.email_back);
+        String email = textEmail.toString();
 
         registerMain = (RegisterMain) getActivity();
 
@@ -46,23 +52,43 @@ public class FragmentEmail extends Fragment {
                 }
 
                 Boolean valid = true;
-                if (textEmail.length() == 0) {
+                if (textEmail.length() == 0 ) {
                     textEmail.setError("Please provide a valid email");
                     valid = false;
 
                 }
+//                if (!Patterns.EMAIL_ADDRESS.matcher(email).matches()) {
+//                    Toast.makeText(getContext(), "Email not valid", Toast.LENGTH_LONG).show();
+//                }
+
                 if (textMobile.length() == 0||textMobile.length()<10) {
                     textMobile.setError("Please provide a valid Mobile number");
                     valid = false;
 
+                }
+                if(textName.length()==0)
+                {
+//            Toast.makeText(getContext(), "F", Toast.LENGTH_SHORT).show();
+                    textName.setError("Name is required");
+                    valid=false;
                 }
                 if(valid) {
                     FragmentTransaction fr = getParentFragmentManager().beginTransaction();
                     fr.replace(R.id.register_container, new FragmentPassword());
 //                fr.addToBackStack(null);
                     fr.commit();
+                    registerMain.myEdit.putString("name", textName.getText().toString());
+                    registerMain.myEdit.commit();
                 }
 
+            }
+        });
+        back.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+
+                Intent intent = new Intent(getActivity(), Login.class);
+                startActivity(intent);
             }
         });
