+++ app/src/main/res/drawable/ic_walk.xml
@@ -1,0 +1,16 @@
+<vector android:height="24dp" android:viewportHeight="512"
+    android:viewportWidth="512" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#FF000000"
+        android:pathData="M312.55,479.9l-56.42,-114 -44.62,-57a72.37,72.37 0,0 1,-10.06 -36.9V143.64H217a40,40 0,0 1,40 40v182.21"
+        android:strokeColor="#000" android:strokeLineCap="round"
+        android:strokeLineJoin="round" android:strokeWidth="32"/>
+    <path android:fillColor="#00000000"
+        android:pathData="M127.38,291.78v-74.07s37,-74.07 74.07,-74.07"
+        android:strokeColor="#000" android:strokeLineCap="round"
+        android:strokeLineJoin="round" android:strokeWidth="32"/>
+    <path android:fillColor="#000000" android:pathData="M368.09,291.78a18.49,18.49 0,0 1,-10.26 -3.11L297.7,250a21.18,21.18 0,0 1,-9.7 -17.79v-23.7a5.65,5.65 0,0 1,8.69 -4.77l81.65,54.11a18.52,18.52 0,0 1,-10.29 33.93zM171.91,493.47a18.5,18.5 0,0 1,-14.83 -7.41c-6.14,-8.18 -4,-17.18 3.7,-25.92l59.95,-74.66a7.41,7.41 0,0 1,10.76 2.06c1.56,2.54 3.38,5.65 5.19,9.09 5.24,9.95 6,16.11 -1.68,25.7 -8,10 -52,67.44 -52,67.44 -2.62,2.98 -7.23,3.7 -11.09,3.7z"/>
+    <path android:fillColor="#FF000000"
+        android:pathData="M257,69.56m-37.04,0a37.04,37.04 0,1 1,74.08 0a37.04,37.04 0,1 1,-74.08 0"
+        android:strokeColor="#000" android:strokeLineCap="round"
+        android:strokeLineJoin="round" android:strokeWidth="16"/>
+</vector>
