+++ app/src/main/res/drawable/ic_water.xml
@@ -1,0 +1,4 @@
+<vector android:height="24dp" android:viewportHeight="512"
+    android:viewportWidth="512" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#FF000000" android:pathData="M265.12,60.12a12,12 0,0 0,-18.23 0C215.23,97.15 112,225.17 112,320c0,88.37 55.64,144 144,144s144,-55.63 144,-144c0,-94.83 -103.23,-222.85 -134.88,-259.88zM272,412a12,12 0,0 1,-11.34 -16,11.89 11.89,0 0,1 11.41,-8A60.06,60.06 0,0 0,332 328.07a11.89,11.89 0,0 1,8 -11.41A12,12 0,0 1,356 328a84.09,84.09 0,0 1,-84 84z"/>
+</vector>
