+++ app/src/main/res/layout/fragment_gender.xml
@@ -136,13 +136,14 @@
         android:id="@+id/gender_next"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_gravity="center_horizontal"
         android:layout_alignParentEnd="true"
         android:layout_alignParentBottom="true"
-        android:layout_marginEnd="80dp"
-        android:layout_marginBottom="80dp"
+        android:layout_marginEnd="65dp"
+        android:layout_marginBottom="20dp"
         android:backgroundTint="@color/primaryColor"
         android:clickable="true"
-        android:gravity="center_horizontal"
+        android:gravity="center"
         android:src="@drawable/ic_baseline_arrow_forward_24"
         android:text="NEXT" />
 
@@ -150,13 +151,13 @@
         android:id="@+id/gender_back"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
+        android:layout_alignParentStart="true"
         android:layout_alignParentBottom="true"
-        android:layout_marginEnd="220dp"
-        android:layout_marginBottom="80dp"
+        android:layout_marginStart="65dp"
+        android:layout_marginBottom="20dp"
         android:backgroundTint="@color/primaryColor"
         android:clickable="true"
-        android:gravity="center_horizontal"
+        android:gravity="center"
         android:src="@drawable/ic_baseline_arrow_forward_24"
         android:text="BACK" />
 </RelativeLayout>
