+++ app/src/main/res/layout/fragment_height.xml
@@ -83,13 +83,14 @@
         android:id="@+id/height_next"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_gravity="center_horizontal"
         android:layout_alignParentEnd="true"
         android:layout_alignParentBottom="true"
-        android:layout_marginEnd="75dp"
-        android:layout_marginBottom="85dp"
+        android:layout_marginEnd="65dp"
+        android:layout_marginBottom="20dp"
         android:backgroundTint="@color/primaryColor"
         android:clickable="true"
-        android:gravity="center_horizontal"
+        android:gravity="center"
         android:src="@drawable/ic_baseline_arrow_forward_24"
         android:text="NEXT" />
 
@@ -97,13 +98,13 @@
         android:id="@+id/height_back"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
+        android:layout_alignParentStart="true"
         android:layout_alignParentBottom="true"
-        android:layout_marginEnd="210dp"
-        android:layout_marginBottom="85dp"
+        android:layout_marginStart="65dp"
+        android:layout_marginBottom="20dp"
         android:backgroundTint="@color/primaryColor"
         android:clickable="true"
-        android:gravity="center_horizontal"
+        android:gravity="center"
         android:src="@drawable/ic_baseline_arrow_forward_24"
         android:text="BACK" />
