--- app/src/main/res/layout/fragment_name.xml
@@ -1,108 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".Register.NameFragment">
-
-    <!-- TODO: Update blank fragment layout -->
-
-
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Hey there!"
-        android:textColor="#131313"
-        android:textSize="20sp"
-        android:textStyle="bold"
-        android:layout_marginLeft="150dp"
-        android:layout_marginTop="120dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.09" />
-
-    <TextView
-        android:id="@+id/textView4"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="200dp"
-        android:text="What's Your Name?"
-        android:textColor="#131313"
-        android:textSize="16sp"
-        android:layout_marginLeft="130dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.498"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView" />
-
-    <com.google.android.material.textfield.TextInputLayout
-        style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="25dp"
-        android:layout_marginRight="25dp"
-        android:hint="Name"
-
-        android:layout_marginTop="250dp"
-        android:textColorHint="@color/black"
-        app:boxStrokeColor="@color/black"
-        app:boxStrokeWidthFocused="2dp"
-        app:hintTextColor="@color/black">
-
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/text_name"
-
-            android:layout_width="match_parent"
-            android:layout_height="66dp"
-            android:textSize="20dp"
-            android:cursorVisible="true"
-            android:inputType="text"
-            android:textColor="@color/black" />
-    </com.google.android.material.textfield.TextInputLayout>
-
-
-
-
-    <com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
-        android:id="@+id/name_next"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="205dp"
-        android:layout_marginTop="400dp"
-        android:backgroundTint="@color/primaryColor"
-        android:clickable="true"
-        android:gravity="center"
-        android:src="@drawable/ic_baseline_arrow_forward_24"
-        android:text="NEXT" />
-
-    <com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
-        android:id="@+id/name_back"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="65dp"
-        android:layout_marginTop="400dp"
-        android:backgroundTint="@color/primaryColor"
-        android:clickable="true"
-        android:gravity="center"
-        android:src="@drawable/ic_baseline_arrow_forward_24"
-        android:text="BACK" />
-
-    <!--        <com.google.android.material.textfield.TextInputEditText-->
-    <!--            android:id="@+id/text_name"-->
-    <!--            android:layout_width="0dp"-->
-    <!--            android:layout_height="61dp"-->
-    <!--            android:cursorVisible="true"-->
-    <!--            android:inputType="text"-->
-    <!--            android:textColor="@color/black"-->
-    <!--            app:layout_constraintBottom_toBottomOf="parent"-->
-    <!--            app:layout_constraintEnd_toEndOf="parent"-->
-    <!--            app:layout_constraintStart_toStartOf="parent"-->
-    <!--            app:layout_constraintTop_toBottomOf="@+id/textView4" />-->
-
-
-
-</FrameLayout>
