+++ app/src/main/res/layout/fragment_password.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
@@ -44,7 +44,7 @@
         android:layout_marginLeft="25dp"
         android:layout_marginRight="25dp"
         android:hint="Password"
-
+        app:passwordToggleEnabled="true"
         android:layout_marginTop="150dp"
         android:textColorHint="@color/black"
         app:boxStrokeColor="@color/black"
@@ -68,7 +68,7 @@
         android:layout_marginLeft="25dp"
         android:layout_marginRight="25dp"
         android:hint="Confirm Password"
-
+        app:passwordToggleEnabled="true"
         android:layout_marginTop="250dp"
         android:textColorHint="@color/black"
         app:boxStrokeColor="@color/black"
@@ -119,11 +119,12 @@
         android:layout_height="wrap_content"
         android:layout_alignParentEnd="true"
         android:layout_alignParentBottom="true"
-        android:layout_marginLeft="205dp"
-        android:layout_marginTop="400dp"
+        android:layout_gravity="center_horizontal"
+        android:layout_marginEnd="65dp"
+        android:layout_marginBottom="20dp"
         android:backgroundTint="@color/primaryColor"
         android:clickable="true"
-        android:gravity="center_horizontal"
+        android:gravity="center"
         android:src="@drawable/ic_baseline_arrow_forward_24"
         android:text="NEXT" />
 
@@ -131,14 +132,13 @@
         android:id="@+id/pass_back"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
+        android:layout_alignParentStart="true"
         android:layout_alignParentBottom="true"
-        android:layout_marginLeft="65dp"
-        android:layout_marginTop="400dp"
+        android:layout_marginStart="65dp"
+        android:layout_marginBottom="20dp"
         android:backgroundTint="@color/primaryColor"
         android:clickable="true"
-        android:gravity="center_horizontal"
+        android:gravity="center"
         android:src="@drawable/ic_baseline_arrow_forward_24"
         android:text="BACK" />
-
-</FrameLayout>
+</RelativeLayout>
