+++ app/src/main/res/layout/fragment_update.xml
@@ -223,7 +223,6 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_centerInParent="true"
-
             android:layout_marginBottom="30dp"
             android:textAlignment="center"
             android:textColor="@color/black"
