+++ app/src/main/res/layout/fragment_weight.xml
@@ -44,10 +44,10 @@
             android:layout_width="93dp"
             android:layout_height="wrap_content"
             android:layout_centerInParent="false"
-            android:layout_marginTop="20dp"
+            android:layout_marginTop="30dp"
             android:layout_marginBottom="30dp"
             android:layout_marginLeft="30dp"
-            android:text="70"
+            android:text="50"
             android:textAlignment="center"
             android:textColor="@color/black"
             android:textSize="25sp"
@@ -61,7 +61,7 @@
             android:layout_height="wrap_content"
             android:layout_centerInParent="true"
             android:layout_marginStart="100dp"
-            android:layout_marginTop="20dp"
+            android:layout_marginTop="30dp"
             android:layout_marginBottom="30dp"
             android:text="KG"
             android:textAlignment="center"
@@ -117,8 +117,9 @@
         android:layout_height="wrap_content"
         android:layout_alignParentEnd="true"
         android:layout_alignParentBottom="true"
-        android:layout_marginEnd="75dp"
-        android:layout_marginBottom="85dp"
+        android:layout_gravity="center_horizontal"
+        android:layout_marginEnd="65dp"
+        android:layout_marginBottom="20dp"
         android:backgroundTint="@color/primaryColor"
         android:clickable="true"
         android:gravity="center"
@@ -129,10 +130,10 @@
         android:id="@+id/weight_back"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
+        android:layout_alignParentStart="true"
         android:layout_alignParentBottom="true"
-        android:layout_marginEnd="215dp"
-        android:layout_marginBottom="85dp"
+        android:layout_marginStart="65dp"
+        android:layout_marginBottom="20dp"
         android:backgroundTint="@color/primaryColor"
         android:clickable="true"
         android:gravity="center"
